requirements.txt
asgiref==3.2.7
astroid==2.4.1
attrs==19.3.0
autopep8==1.5.2
coverage==5.1
dj-database-url==0.5.0
Django==3.0.6
django-on-heroku==1.0.1
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.3.0
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
pycodestyle==2.6.0
pylint==2.5.2
pyparsing==2.4.7
pytest==5.4.2
pytest-django==3.9.0
pytz==2020.1
rope==0.17.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wcwidth==0.1.9
whitenoise==5.1.0
wrapt==1.12.1
